+++ lib/src/components/components.dart
@@ -16,6 +16,7 @@ export 'dialog/foss_dialog.dart';
 export 'drawer/foss_drawer.dart';
 export 'meter/foss_meter.dart';
 export 'number_field/foss_number_field.dart';
+export 'otp_field/foss_otp_field.dart';
 export 'popover/foss_popover.dart';
 export 'progress/foss_progress.dart';
 export 'radio/foss_radio.dart';